backend/auth-mw/requirements.txt
uvicorn[standard]
fastapi
datetime
pytest
gunicorn
pydantic==1.10.8
sqlalchemy
psycopg2-binary
passlib[bcrypt]
pyjwt
pydantic[email]